Remove namespace pollution via plugins
Inside of using the Sinuba namespace, put plugins in the
Sinuba::SinubaPlugins namespace.  Also move the base plugin from
Sinuba to Sinuba::SinubaPlugins::Base.  I'll miss the meta-ness
of Sinuba being isn't own plugin, but it's best to be consistent.

With this change, all constants in the Sinuba namespace are
prefixed with Sinuba, which should avoid any conflicts when
subclassing and using plugins.