Add r.is method for exact matches
This adds a :term matcher which matches only if the path is empty,
which can use used as the last matcher on block to ensure the
path matches exactly. r.is just adds this matcher before calling
r.on, to make it easy to use.

While here, remove the :root matcher.  It's a bad idea to treat
an empty path as a root path, the root path can be handled by
"", and the empty path by :term.