Add render_opts[:ext]
Previously, opts[:engine] was pulling double duty as both the
extension to use for template paths, and as the engine used
by Tilt for inline templates.  This splits the options to the
behavior can be different, but falls back to having the
template path using :engine if :ext is not set.