Make r.get/r.post call r.is if arguments given
While it can be considered confusing that the methods dispatch
different based on the number of arguments, this actually does
what you want in almost all cases.  In general, you would call
r.get without arguments in two cases.  The first case is when
your route hierarchy begins with the request method:

  route do |r|
    r.get do
    end
    r.post do
    end
  end

In this case, you want to call r.on, as the path checking is
done later.

The second case is when your route hierarchy ends with the
request method:

  route do |r|
    r.is "foo/:id" do
      r.get do
      end
      r.post do
      end
    end
  end

In this case, you already fully matched the patch, so using
r.is is redundant.

If you provide arguments to r.get or r.post, it's likely
because you are giving it a full path and want to route on
the full path, similar to how Sinatra works:

  route do |r|
    r.get "foo/:id" do
    end
    r.post "bar/:id" do
    end
  end

In this case, you want to fully match the path and the request
method.

The only time this doesn't do what you want is if you want
to partially match the path as well as match the request
method.  You can use the :method hash matcher for that case.