Support a load_dependencies method for plugins
This is needed if a plugin depends on another plugin, to insure
that that plugin is loaded before the other plugin modules
are included.  You can't do this inside configure, as that would
load the plugin dependencies after the current plugin, meaning
you couldn't call override a method in a plugin and call super
to get the method in the dependency.