Add hooks plugin for before/after hooks
before hooks are not called with anything, since they can access
the request and response via methods.

after hooks are called with the raw rack response array of
status, headers, body, and can mutate the raw array to
change the response.  They can't use the response method to
manipulate it, unless they want to call
Array#replace(response.finish) on the block argument.

If an exception is raised, after will still be called, but
the response will be nil, and the exception will be available in
$!.