Allow specifying :callback and :loop options to stream
This switches stream and Stream#initialize to use a hash based API.
Move keep_open to the :keep_open option, and add a :scheduler option
for easy overriding of the scheduler.

Add a :callback option to allow specifying an initial callback.
Add a :loop option that automatically loops while the stream hasn't
been closed.

With these two options, streaming code is a lot DRYer.