Make default_headers plugin have class level access to headers
This involves storing them in the class, and having the response
class look them up at runtime, instead of just defining a method
in the response class that returns a dup of the hash.

This also makes it so you can load the plugin multiple times
into the same app.