Optimize r.is terminal matcher, remove :term hash matcher
The terminal matcher added by r.is is very common, so it deserves
an optized implementation.  This makes RodaRequest::TERM an empty
frozen object, which will only match itself in a case statement,
and adds an additional case to r.match.

While the term matcher was documented in the README, it wasn't
actually used in the specs, and it's unlikely someone is using
it directly.