Add per_thread_caching plugin, for using separate caches per thread instead of shared thread-safe caches
The shared thread-safe cache may save memory, but it may not
perform as well as a per-thread cache due to the locking.
I'm not positive it performs better, that can depend on many
things, such as CPU cache sizes.  Still, it's good to have an
option for JRuby/Rubinius users if the mutex used by the default
thread-safe cache is a bottleneck.