Change r.root to only match get requests
In general, non-GET requests to root are rare, so r.root with
no arguments should only allow get requests.  It would be possible
to still accept an optional method argument and just make get
the default, but I think just having people do r.post '' for posting
to the root is fine.