Optimize is and verb match blocks with no arguments
This removes r._on and splits it into two methods, r.always and
r.if_match.  r.always takes no arguments and always calls the
block, yielding nothing. r.if_match takes an array of matchers,
only calls the block if all matchers match, and yields the
captures.

r.is now checks if the arguments are empty.  If so, it does an
immediate check of the current path, and calls always if the path
is empty.  Otherwise, it adds the terminal condition and calls
if_match.

r.get and r.post now call always for empty args, and otherwise
add the terminal matcher and call if_match.

While here, fix RDoc for some request methods to reflect reality.