Fix caching of extension matcher
Previously, it would conflict with a string/symbol matcher due
to the same key being used.  Currently, just use an array with
the first element being the symbol :extension.  This could be
optimized so that it doesn't need to allocate an array, but
it doesn't seem worth it as the :extension matcher isn't that
common.