Add not_allowed plugin, for automatically returning 405 Method Not Allowed responses
This refactors the core so that RequestClassMethods contains a
def_verb_method method that will create a verb method in the given
module.  Because this is an instance method in the
RequestClassMethods, you have to wait until after the Base plugin
has been loaded for it to define the get and post methods.

This simplifies the all_verbs module, since that can just call
def_verb_method.