Add optimized is_get? method
The default get? method allocates a couple of strings, save those
allocations by using frozen string constants. As GET is the most
common request method, it makes sense to optimize for it.

This also allows is_get? to be overridden for different behavior,
which is dangerous to do with get? as that is defined the
superclass.