Use better example on the why page
Looking back at some of the response to Roda's release, it seems
the main code example I used was too complex and made it difficult
for people to see the advantages.

This new example hopefully does a better job of how Roda DRYs up
code without some of the negative issues that the previous example
has, and is closer to how Roda has DRYed up code in the Sinatra
apps I've converted to Roda.

Also, because some people wanted explicit performance benefits,
go over the general issues with large apps, and link to a benchmark
for small apps.