Add error_email plugin, for easily emailing an error notification for an exception
As it says in the documentation, this is only appropriate for
low traffic sites, but on my low traffic sites I'm usually doing
something like this.

The exception_notification gem handles similar use cases, but it
is heavyweight in terms of dependencies. This uses the net/smtp
library from stdlib to avoid an external dependency.