Pick up newly added named routes while running in the multi_route plugin, useful for development
If you are using a development environment reloader that picks
up new files and loads them into the current process, and you
added a file with a new named route, and you were using
multi_route to dispatch to your named routes, then the route
would not be picked up.  This clears the named route regexp
when a route is added, so a new one will be generated on the next
call to multi_route, allowing it to pick up the new route.