Make r.multi_route prefer longer route if multiple routes have the same prefix
Previously, if you had named routes "foo" and "foo/bar", and you
got a request for "/foo/bar/baz" the route chosen would depend on
the order the entries were added to the hash (and would be
arbitrary on ruby 1.8).  This sorts the entries and reverses the
output, ensuring that longer entries come before shorter entries
in the regexp union, so that they are matched first.