Use allocate instead of new in rack app
Roda by default does not define initialize for the instance scope,
so calling new is the same as calling allocate, just slower.

It may be faster to use new(env).call(&block) if #call doesn't
call #_route, but that would change the API, and in my testing it
isn't conclusively faster in all cases.