Add response.finish_with_body
This should be used whenever you want to keep the current request
status and headers, but want to use a different body (that may not
just be a simple array).

Since we now support using a different response body, it doesn't
make sense to automatically set a Content-Length header that may
not match.  So change things so that response.write does not
set a Content-Length header, just update the length.  Have
response.finish set the Content-Length header to the length.
Don't have response.finish_with_body set Content-Length.
This also has the advantage of being faster if calling res.write
multiple times, since the Content-Length header will only be
set once.

Update the chunked and streaming plugins to use this new method.