Make InstanceMethods in plugins not include constants, as they would pollute the constant namespace
The base, chunked, flash, and render_each plugins all used
constants inside InstanceMethods, which resulted in those constants
becoming constants in the Roda class that loaded the plugin. This
breaks one of the primary design philosophies of Roda, which is
that it does not pollute your namespaces.

Fix this by moving the constants lexically up to the plugin module.
The plugin module itself is not included into the Roda class, so
this will not pollute namespace.  Because constants are resolved
lexically before using the inheritance hierarchy, this should be
backwards compatible.

This removes the EMPTY_STRING constant from the render_each plugin,
since it wasn't used.