Add mailer plugin, allowing your Roda application to send emails
While this can be used by the same app that handles web requests,
that's probably a bad idea.  However, the advantages the routing
trees offer are true for email applications just as they are for
web application.  Plus, this allows you to use Roda's render
plugin to create email response bodies, and can share code/helpers
with the web app.