Add delegate plugin, for easily creating methods that delegate to request or response
One of Roda's main advantages over other ruby web frameworks is
that it does not pollute namespaces.  For the method namespace, it
does this by adding methods to the request or response objects
instead of the route block scope object, so that such methods
will not conflict with methods the user wants to use.

However, in some cases it makes for a simpler and cleaner DSL to
not have explicitly reference the request and response instances
when calling such methods.  The delegate plugin allows you to do
this.  So the user can easily pollute their own application's
namespace if they want to, without polluting the namespace for
other users.