Make RodaRequest.consume_pattern not capture pattern by default
Previously, consume_pattern added a capture for the pattern
in order to update SCRIPT_NAME.  However, that information
can be extracted from post_match data, since we are storing
the initial path, so it isn't necessary.

This patch actually slows down the default matching by about
2%, but if the static_path_info plugin is used, this speeds
up matching by about 3.5%.  With this patch, Roda with the
static_path_info plugin is about 18% faster when matching,
compared to the default behavior without this patch.

Note that this can break backwards compatibility for users
that were manually calling RodaRequest#consume and capturing
the pattern.  Such users need to change their code to not
capture the pattern.