Fix handling of multipart content types in the mailer plugin
Previously, if the mail plugin created a multipart mail, it would
break as the Content-Type would be replaced.  This changes things
so that the default Content-Type header is deleted before
processing the routing block for mail routes.  Also, when finishing
the mail request, the Content-Type header is removed from the
headers to automatically add to the message.  The header (or the
default :content_type setting) is then used to set the content_type
for the message using the mail API, making sure to set the content
type only for the specific (non-attached) part in a multiple/mixed
message.  A multipart/alternative message does not have a content
type modified, as text_part/html_part are used to specify the
content parts for those.