Add delay_build plugin, to delay building the rack app until Roda.app is called
This is only needed in extreme cases, where hundreds of middlware
are added after the route block has been defined.  However, in
those cases, every time a middleware is added, it rebuilds the app,
which is an O(n) operation on the number of middleware, which
means the total time is O(n^2).  By not rebuilding unless build! is
called, this turns into an O(n) operation.