Describe Roda as a toolkit
Depending on the plugins you load when Roda, the resulting code for
the web application can be significantly different.  So Roda is
best described as a toolkit that can be used to build routing tree
web frameworks.

While here, change the RDoc to point to the website instead of GitHub
as the homepage.