Organize plugins on documentation page into sections
This should make it easier than going through a long list.

Stop documenting all plugins in the README, just have it link
to the website.

Add links to external libraries for gems that are related to
Roda, and another external plugin.

Add id tags to make it easy to link to specific parts of the
documentation page.