Support :add_script_name, :name, :url, and :url_only options when creating named paths in the path plugin
Unfortunately, due to how Roda works by default, :add_script_name
is going to generate methods that give different results depending
on where they are called in the routing tree.  You can use the
static_path_info plugin to fix that (and it will be fixed by default
in Roda 2).