Remove RodaRequest.def_verb_method
Instead of using def_verb_method to define methods like get and post,
just define them the usual way.

In the all verbs plugin, inline the previous call to def_verb_method,
and define the methods in the module's RequestMethods, instead of inside
a request module specific to each application that loads the plugin.

Have the not_allowed plugin depend on the all_verbs plugin to work
around load order issues.  Instead of calling def_verb_method to rewrite
the methods, just override each of the verb methods and call super to
get the default behavior.

Extract a method_not_allowed request private method out to DRY up some
code in the not_allowed plugin.

Fix an bug in the not_allowed plugin when calling a verb method outside
of an is block when an is block has already been called.  You could
probably only hit this bug when using the pass plugin.