Add path_to_match and matched_path request methods
With the default Roda behavior, these are just aliases for
PATH_INFO and SCRIPT_NAME.  But when static_path_info is used
you want to provide the same information.  path_to_match
already existed, but was private, so make it public.  matched_path
in static path info is pretty simple, just adding the already
matched part of PATH_INFO to SCRIPT_NAME.