Don't override explicit nil :default_encoding template option in the render plugin
Some combinations of tilt/ruby have an issue where setting a
default encoding can lead to a frozen string error.  I don't
plan to work around the bug inside Roda, but one way of doing
so outside of Roda is to just not set a default_encoding, which
this allows you to do.