Make RodaResponse a subclass of Object instead of Rack::Response
This was a bug added when Roda was forked from Cuba.  RodaResponse
doesn't use any of the Rack::Response features, so there is no
reason it should be a subclass.

Response#body was used implicitly in the chunked plugin, so add
a method for that.  Note that the result should be treated as
read only, modifying the body directly can break things.

This potentially breaks backwards compatibility if users were
calling Rack::Response methods that RodaResponse did not override.
In many cases, those methods would raise an error, but possibly
not all.