Make not_found plugin clear headers for response if it is not found
This fixes an issue where using not_found would generally result
in an incorrect Content-Length if the body was written to inside
the not_found block (which is the usual case).

I debated whether to just clear the Content-Length header, but
for similar reasons as the error_handler plugin using a new
response, I think it's best to clear all the headers.

This can cause backwards compatibility issues if you relied
on the previous behavior of setting headers in the route block
and having them preserved in the not_found block.  If someone
is actually relying on that, I'm opening to adding an option
that just clears specific headers.