Move Roda.hash_matcher to hash_matcher plugin
Adding custom hash matchers is probably not done in most apps,
so using a plugin for it is more suitable than having it in base.

This slightly changes hash_matcher so that it adds the method
directly to the class instead of a module included in the class
(so it no longer need to rely on request_module), but that
shouldn't have an effect.