Move :param and :param! hash matchers to the param_matchers plugin
While these are more common than the extension matcher, they still
aren't all that common.  One reason for that is that if you match
on a param without a fallback, you end up with a 404 response, when
the page actually exists, which doesn't seem right.  I think they
are more suitable for a plugin than base.