Add Roda#render_template private method for template rendering, for use by plugins
Shortly, there will be a plugin added that overrides render to use
a layout, so add an alias for the default render behavior, and have
all plugins that want to render without a layout use that method.

Marking it private is mostly to indicate it is for plugin use and
not for applications to use directly.  However, since the Roda
instance is self when evaluating the route block and rendering
templates, this doesn't prevent applications from doing so.