Have error_handler rescue ScriptError instead of SyntaxError
ScriptError is the superclass of SyntaxError, so this still catches
SyntaxError, but also LoadError (raised by require) and
NotImplementedError (raised by TSort).  Of the standard ruby
exception classes, this doesn't rescue: Exception, NoMemoryError,
SignalException, SystemExit, and fatal.

I originally was considering making it possible for a user to
pick the exception classes, but that's probably overkill with
this change.