Speed up Roda#request and #response
attr_reader defined methods are faster than regular methods. But
because the instance variables are prefixed with an underscore,
but the methods are not, attr_reader won't work directly.

Define the methods using attr_reader with the underscore, then
alias to the ununderscored version, then remove the underscored
version.  This makes the methods faster.