add_file in the mailer plugin now adds the files after the email body instead of before
This fixes some issues where an empty body would be used, even if
the should set a body.  IMO, this is a bug in mail's API, where
add_file before body doesn't work as expected, but it's easier
to work around the issue just by delaying the add_file calls
until finish.