Make :layout=>false option in render plugin override previous layout template
Previously, loading the plugin multiple times like this:

  plugin :render, :layout=>'a'
  plugin :render, :layout=>nil

Resulted in the default layout template being 'a', even though
a layout would not be used by default.  Now, it resets the layout
such that there is no default layout template.  If you want to
not use a layout by default, but still have a default layout
template, you need to use :layout_opts=>{:template=>'a'}.

This code makes the render plugin more similar to the assets
plugin in that the original options used are now stored separately.
This will be used by some upcoming features.