Recognize Roda :root option in render and assets plugins
This makes the :root option the default base filesystem path for
the assets and render plugins.  So any relative filesystem paths
used will be relative to :root.

For both the assets and render plugins, Roda now does the
following:

1) If the option for the path is not given, Roda will use the
default path relative to the :root option.

2) If a relative path is given for the option, Roda will make it
relative to the :root option.

3) If an absolute path is given for the option, Roda will use
it directly.

In all cases, the resulting path used will now be made absolute
using File.expand_path.