Support :add_script_name app option in assets and path plugins
This removes the just add support for the assets plugin :script_name
option, into a better named and more appropriate :add_script_name
app option.  It also changes the path plugin so that the default
for the :add_script_name option in the path plugin is the
application's add_script_name.