Add support for registering classes in the path plugin for use with Roda#path
This can simplify some usage of paths.  Instead of creating separate
*_path instance methods for each type of path, it just adds a single
path instance method, which dispatches to different blocks based on
the class of the first argument. So instead of:

  bar_path(Bar_instance)

which is a little redundant, you can do:

  path(Bar_instance)

The only thing you have to do is register the path slightly differently:

  app.path(Bar){|bar|...}

instead of:

  app.path(:bar){|bar| ...}