Merge :locals set in the render plugin options into :locals provided in call to view/render
Previously, if you provided :locals in the render plugin options,
and also provided locals as the view/render option, it would
ignore the plugin option.  This broke the principle of least
surprise, since the plugin options should provide defaults, and
unless a specific local variable was overridden in the call,
the default should be used.

You can set locals both for regular renders and for layout renders,
and they should have separate locals hashes, so introduce an
entry in the :layout_opts plugin hash that specifies that the
options are for the layout.  Use that entry to determine whether
the template locals should be merged.  The layout locals are
always merged when rendering a layout.