Rename view_subdirs plugin to view_options, add support for branch/route specific view/layout options/locals
This can make it easier to deal with branches or routes that want
to use options different than the default plugin options.
Previously, if you wanted something different, you had to pass
options to every call to render/view where you wanted to change
behavior.  That's not very DRY, and sort of defeats one of the
main advantages of using a routing tree.

With this plugin, you can set the options/locals to use for a
given branch, and they will take precedence over the render
plugin options, but they have lower precedence than options
given directly to a view/render method.

As view_options is a superset of view_subdirs, and attempts to
load the view_subdirs plugin now end up loading view_options,
there should be no effect on backwards compatibility.