Have add_file in the mailer plugin support blocks, which are called after the file has been added
Mail's API does not offer a way to add an attachment and set its
content type at the same time.  The add_file message takes a hash,
but only respects the :filename and :content entries, it doesn't
pass that hash to attachements[]=.  You can't call attachements[]=
manually, as that won't automatically convert the mail to multipart.

This provides a workaround so that you can add a file and set the
content type for it (or perform other operations on it).  Hopefully
the Mail API can be fixed so this will be simpler in the future.