Update README and website to better explain Roda
This adds a section for Roda's goals near the top of the README
and why page.  It also updates the first example to show how
variables are set at the branch level and can be used in the
routes underneath that branch.

Some other minor wording changes and fixes while here.