Support exceptions when automatically escaping <%= %> tags in render plugin
Previously, the choice when escaping ERb templates was all or
nothing.  Now you can specify a :escape_safe_classes option
which will be an array of classes not to escape.  This is useful
if you have helpers or other libraries that return already escaped
strings and you want to have <%= %> not double escape them.

For complete control, you can specify an :escaper option, which
will be used for escaping <%= %> strings.