Add json_parser plugin, for parsing request bodies in JSON format
Previously, I had recommended the rack-parser middleware for this.
However, because that is middleware, it will parse an uploaded
request body for every request.  With this plugin approach, it will
only parse the request body if you call r.params or r.POST.

This plugin should also be faster than the rack-parser approach,
especially for routes where the params are not loaded.