Avoid rehashing hashes at runtime in plugins
From tenderlove I found out that Hash#dup is slower than Hash[],
and from my own benchmarking I've found that Hash#merge is slower
than Hash[].merge!.  This is because both Hash#dup and Hash#merge
rehash the hash.

Hash[] isn't exactly the same as Hash#dup, and Hash[].merge! is
not exactly the same as Hash#merge, as Hash[] doesn't copy the
default value/proc or any instance variables set on the hash.
But these changes are for options or header hashes, which
shouldn't have either of those issues.