The render plugin and render/view :ext option is now replaced by the :engine option
There's no reason there should be two options.  The way
Tilt::Mapping works, engine names are file extensions, so the two
should be synonomous.  Have the :engine option take precedence,
but still support the :ext option if the :engine option is not
specified.

This simplifies template_path by making sure that the template
engine and views directory are specified in opts, so it doesn't
need to look in render_opts.

This does change one of the specs, but I don't think the spec
made much sense, since 'erb' was never specified as an engine,
and it seems unlikely users would want to default to str file
templates but erb inline templates.

While here, make sure the :engine and :views options are frozen.