Optimize template cache lookup in render plugin when :cache_key is given
If :cache_key is given explicitly, then a lot of the logic inside
find_template can be skipped, since it is either used to determine
the cache key to use, or it is cached inside the template.

One exception to this is the :locals handling, which needs to be
handled specially, so move that into a merge_render_locals method
that is always called when rendering.

Additionally, split a retrieve_template method out of the render
method, allowing future code to get access to the underlying
template objects.

This change can speed up rendering of simple templates by 40-50%.
It also adds some documentation showing how to easily speed up
rendering of templates across your entire application.