Add :include_request option for json plugins
When using the json plugin, sometimes we want to serialize the object
differently depending on the request headers or request parameters. For
example, we might want to give the client the ability to specify which
post associations they want returned, with a request like
/posts?include=author,comments.

If :include_request=>true is passed to the request, the request object
will be passed as the second argument when calling the serializer. We
add it to the json_parser plugin as well for consistency.