Add websockets plugin, for integration with faye-websocket
The specs use thin for testing, since that was easy to setup
and worked correctly.  I tried to get the specs working with
puma, but unfortunately I was not able to do so.

Thin doesn't install on JRuby, so I didn't put the thin dependency
in the gemspec. In the travis gemfile install thin except when on
JRuby, so that Travis should work correctly.  This will result in
not testing the websockets plugin on JRuby, unfortunately.