Merge options when loading csrf plugin multiple times
Previously, this would load the middleware multiple times with
separate options, now it only loads the middleware once and
merges the new options into the previous options.

This allows plugins that depend on csrf to be loaded, while also
allowing the application to use the csrf plugin with specific
options.