Alias #<< to #write for streaming IO compatibility
The standard convention of writable streams is that they respond to
`#write`. If we make the stream object respond to `#write`, we could for
example use `IO.copy_stream` to efficiently copy a stream to it.

```rb
stream do |out|
  IO.copy_stream(StringIO.new(content), out)
end
```

We also make `#write` the main method and the operator its alias.