Allow error handler access to the request's remaining_path
I think having access to the remaining_path would be useful
in the error handler, so it can know which branch of the
routing tree was already taken by comparing it to PATH_INFO.

This removes the ensure block in if_match, only resetting the
remaining path if the block doesn't match.  If the block
matches, there is no reason to reset the remaining path, since
the block should be halting.  The only cases where this isn't
true are when using the pass and class_level_routing plugin,
which can be modified to reset the remaining path.

Note this can break backwards compatibility with external code
that uses non-local jumps similar to the pass plugin (via rescue
or call), or does processing after the route block similar
to the class_level_routing plugin.